feat: 식물 등록 기능 개발 중

+++ src/main/java/com/example/plantappbackend/service/PlantService.java
@@ -1,7 +1,7 @@
 package com.example.plantappbackend.service;
 
-import com.example.plantappbackend.Repository.PlantRepository;
 import com.example.plantappbackend.model.Plant;
+import com.example.plantappbackend.repository.PlantRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,7 +13,7 @@ public class PlantService {
     @Autowired
     private PlantRepository plantRepository;
 
-    public List<Plant> getPlantsByUser(Long userId) {
+    public List<Plant> getPlantsByUser(int userId) {
         return plantRepository.findByUserId(userId);
     }
